Create README.md
This is a beginner html project. I know how to use html and css to make more complex stuff but this is good practice for learning how to set up projects while keeping track of changes. 
I'm also practicing using bash to create and manage the projects that I upload here.